+++ lib/src/models/line.dart
@@ -20,7 +20,7 @@ class Line extends Vector {
     LayoutAlign layoutAlign,
     LayoutConstraint constraints,
     String transitionNodeID,
-    int transitionDuration,
+    double transitionDuration,
     EasingType transitionEasing,
     double opacity,
     SizeRectangle absoluteBoundingBox,
