+++ lib/src/models/rectangle.dart
@@ -28,7 +28,7 @@ class Rectangle extends Vector {
     LayoutAlign layoutAlign,
     LayoutConstraint constraints,
     String transitionNodeID,
-    int transitionDuration,
+    double transitionDuration,
     EasingType transitionEasing,
     double opacity,
     SizeRectangle absoluteBoundingBox,
