+++ lib/src/models/text.dart
@@ -37,7 +37,7 @@ class Text extends Vector {
     LayoutAlign layoutAlign,
     LayoutConstraint constraints,
     String transitionNodeID,
-    int transitionDuration,
+    double transitionDuration,
     EasingType transitionEasing,
     double opacity,
     SizeRectangle absoluteBoundingBox,
